+++ libraries/common/src/test/java/androidx/media3/common/util/UtilTest.java
@@ -1314,9 +1314,35 @@ public class UtilTest {
   }
 
   @Test
-  public void getStringForTime_withNegativeTime_setsNegativePrefix() {
-    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ -35000))
-        .isEqualTo("-00:35");
+  public void getStringForTime_withRangeOfValues() {
+    assertThat(getStringForTime(C.TIME_UNSET)).isEqualTo("00:00");
+    assertThat(getStringForTime(0)).isEqualTo("00:00");
+    assertThat(getStringForTime(413)).isEqualTo("00:00");
+    assertThat(getStringForTime(800)).isEqualTo("00:01");
+    assertThat(getStringForTime(10_000)).isEqualTo("00:10");
+    assertThat(getStringForTime(65_000)).isEqualTo("01:05");
+    assertThat(getStringForTime(3_661_000)).isEqualTo("1:01:01");
+    assertThat(getStringForTime(-4_000)).isEqualTo("-00:04");
+  }
+
+  @Test
+  public void getStringForTime_withFormatter_withRangeOfValues() {
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ C.TIME_UNSET))
+        .isEqualTo("00:00");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ 0))
+        .isEqualTo("00:00");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ 413))
+        .isEqualTo("00:00");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ 800))
+        .isEqualTo("00:01");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ 10_000))
+        .isEqualTo("00:10");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ 65_000))
+        .isEqualTo("01:05");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ 3_661_000))
+        .isEqualTo("1:01:01");
+    assertThat(getStringForTime(new StringBuilder(), new Formatter(), /* timeMs= */ -4_000))
+        .isEqualTo("-00:04");
   }
 
   @Test