+++ libraries/ui_compose/src/test/java/androidx/media3/ui/compose/utils/TestPlayer.kt
@@ -21,25 +21,33 @@ import androidx.media3.common.PlaybackParameters
 import androidx.media3.common.Player
 import androidx.media3.common.SimpleBasePlayer
 import androidx.media3.common.VideoSize
-import com.google.common.collect.ImmutableList
+import androidx.media3.common.util.Assertions.checkState
+import androidx.media3.common.util.Util.msToUs
 import com.google.common.util.concurrent.Futures
 import com.google.common.util.concurrent.ListenableFuture
+import kotlinx.coroutines.test.TestCoroutineScheduler
 
 /**
  * A fake [Player] that uses [SimpleBasePlayer]'s minimal number of default methods implementations
  * to build upon to simulate realistic playback scenarios for testing.
  */
-internal class TestPlayer : SimpleBasePlayer(Looper.myLooper()!!) {
+internal class TestPlayer(
+  playbackState: @Player.State Int = STATE_READY,
+  playWhenReady: Boolean = false,
+  playlist: List<MediaItemData> =
+    listOf(
+      MediaItemData.Builder(/* uid= */ "First").setDurationUs(1_000_000L).build(),
+      MediaItemData.Builder(/* uid= */ "Second").setDurationUs(2_000_000L).build(),
+    ),
+  playbackSpeed: Float = 1f,
+) : SimpleBasePlayer(Looper.myLooper()!!) {
   private var state =
     State.Builder()
       .setAvailableCommands(Player.Commands.Builder().addAllCommands().build())
-      .setPlaylist(
-        ImmutableList.of(
-          MediaItemData.Builder(/* uid= */ "First").setDurationUs(1_000_000L).build(),
-          MediaItemData.Builder(/* uid= */ "Second").setDurationUs(2_000_000L).build(),
-        )
-      )
-      .setPlayWhenReady(true, PLAY_WHEN_READY_CHANGE_REASON_USER_REQUEST)
+      .setPlaybackState(playbackState)
+      .setPlayWhenReady(playWhenReady, PLAY_WHEN_READY_CHANGE_REASON_USER_REQUEST)
+      .setPlaylist(playlist)
+      .setPlaybackParameters(PlaybackParameters(playbackSpeed))
       .build()
 
   var videoOutput: Any? = null
@@ -81,7 +89,7 @@ internal class TestPlayer : SimpleBasePlayer(Looper.myLooper()!!) {
         .setContentPositionMs(positionMs)
         .build()
     if (mediaItemIndex == state.playlist.size - 1) {
-      removeCommands(Player.COMMAND_SEEK_TO_NEXT)
+      removeCommands(COMMAND_SEEK_TO_NEXT)
     }
     return Futures.immediateVoidFuture()
   }
@@ -122,6 +130,47 @@ internal class TestPlayer : SimpleBasePlayer(Looper.myLooper()!!) {
 
   fun setPosition(positionMs: Long) {
     state = state.buildUpon().setContentPositionMs(positionMs).build()
+    invalidateState()
+  }
+
+  /**
+   * Sets the [PositionSupplier] for the current content playback position using extrapolation.
+   * Start and elapsed time are measured using the [TestCoroutineScheduler], and playback speed is
+   * taken into account.
+   *
+   * Note, that this approach only works
+   * * for a single media item (does not consider current item's duration and playlist length)
+   * * assuming the Player is playing (does not react to pauses and discontinuities)
+   * * assuming the playback speed does not change after the supplier is set
+   */
+  fun setPositionSupplierDrivenBy(testScheduler: TestCoroutineScheduler) {
+    checkState(state.playlist.size == 1, "Playlist must contain exactly one item")
+    checkState(isPlaying, "Player must be playing")
+    val startTime = testScheduler.timeSource.markNow()
+    val currentPositionMs = currentPosition
+    val positionSupplier = {
+      val elapsedTime = testScheduler.timeSource.markNow().minus(startTime)
+      currentPositionMs + (elapsedTime * playbackParameters.speed.toDouble()).inWholeMilliseconds
+    }
+    state = state.buildUpon().setContentPositionMs(positionSupplier).build()
+    invalidateState()
+  }
+
+  fun setBufferedPositionMs(bufferedPositionMs: Long) {
+    state = state.buildUpon().setContentBufferedPositionMs { bufferedPositionMs }.build()
+    invalidateState()
+  }
+
+  fun setDuration(uid: String, durationMs: Long) {
+    val index = state.playlist.indexOfFirst { it.uid == uid }
+    if (index == -1) {
+      throw IllegalArgumentException("Playlist does not contain item with uid: $uid")
+    }
+    val modifiedPlaylist = buildList {
+      addAll(state.playlist)
+      set(index, state.playlist[index].buildUpon().setDurationUs(msToUs(durationMs)).build())
+    }
+    state = state.buildUpon().setPlaylist(modifiedPlaylist).build()
     invalidateState()
   }